backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
channels
channels-redis
Pillow
pyotp
qrcode
requests
gunicorn
whitenoise
dj-database-url